event: This now supports Remote Events.
Basically this means a new class called RemoteManager
which listens by default on 0.0.0.0:64000 on the UDP
protocol for nodes to send it events.

Event are processed locally first, then sent over
the remote channel. In essence, this creats a
kind of a mesh event network where RemoteManager's
are able to send and receive events to each other.

In theory it would be possible to build distributed
apps with this, and build apps whoose parts need
not run on the same machine. (ie: Agents)

committer: James Mills <prologic@shortcircuit.net.au>