event: Enabled broadcast. event: Send remote events first, then local. Return for local events.
NOTE: With broadcast events, please be _very_ careful.
You must be careful not to have a situation where:
A sends an event to broadcast nodes A, B, C, D and
A processes the same event it sent out. An infinite
loop of events will occur between the nodes.

committer: James Mills <prologic@shortcircuit.net.au>