* Fixed irc, sockets, timers and examples.
I broke a lot of code by removing the keyword argument
from push and send in the EventManager. I thought
that since I've never used them, they can't really be
useful for anything. And it was causing pickling problems
with the RemoteManager when source as an instnace of
EventManager. :/

committer: James Mills <prologic@shortcircuit.net.au>